TST/REF: duplicated hist tests
