+++ pandas/tests/plotting/test_series.py
@@ -365,147 +365,6 @@ class TestSeriesPlots(TestPlotBase):
         result = [x.get_text() for x in ax.texts]
         assert result == expected
 
-    def test_hist_df_kwargs(self):
-        df = DataFrame(np.random.randn(10, 2))
-        _, ax = self.plt.subplots()
-        ax = df.plot.hist(bins=5, ax=ax)
-        assert len(ax.patches) == 10
-
-    def test_hist_df_with_nonnumerics(self):
-        # GH 9853
-        with tm.RNGContext(1):
-            df = DataFrame(np.random.randn(10, 4), columns=["A", "B", "C", "D"])
-        df["E"] = ["x", "y"] * 5
-        _, ax = self.plt.subplots()
-        ax = df.plot.hist(bins=5, ax=ax)
-        assert len(ax.patches) == 20
-
-        _, ax = self.plt.subplots()
-        ax = df.plot.hist(ax=ax)  # bins=10
-        assert len(ax.patches) == 40
-
-    def test_hist_legacy(self):
-        _check_plot_works(self.ts.hist)
-        _check_plot_works(self.ts.hist, grid=False)
-        _check_plot_works(self.ts.hist, figsize=(8, 10))
-        # _check_plot_works adds an ax so catch warning. see GH #13188
-        with tm.assert_produces_warning(UserWarning):
-            _check_plot_works(self.ts.hist, by=self.ts.index.month)
-        with tm.assert_produces_warning(UserWarning):
-            _check_plot_works(self.ts.hist, by=self.ts.index.month, bins=5)
-
-        fig, ax = self.plt.subplots(1, 1)
-        _check_plot_works(self.ts.hist, ax=ax)
-        _check_plot_works(self.ts.hist, ax=ax, figure=fig)
-        _check_plot_works(self.ts.hist, figure=fig)
-        tm.close()
-
-        fig, (ax1, ax2) = self.plt.subplots(1, 2)
-        _check_plot_works(self.ts.hist, figure=fig, ax=ax1)
-        _check_plot_works(self.ts.hist, figure=fig, ax=ax2)
-
-        with pytest.raises(ValueError):
-            self.ts.hist(by=self.ts.index, figure=fig)
-
-    def test_hist_bins_legacy(self):
-        df = DataFrame(np.random.randn(10, 2))
-        ax = df.hist(bins=2)[0][0]
-        assert len(ax.patches) == 2
-
-    def test_hist_layout(self):
-        df = self.hist_df
-        with pytest.raises(ValueError):
-            df.height.hist(layout=(1, 1))
-
-        with pytest.raises(ValueError):
-            df.height.hist(layout=[1, 1])
-
-    def test_hist_layout_with_by(self):
-        df = self.hist_df
-
-        # _check_plot_works adds an ax so catch warning. see GH #13188
-        with tm.assert_produces_warning(UserWarning):
-            axes = _check_plot_works(df.height.hist, by=df.gender, layout=(2, 1))
-        self._check_axes_shape(axes, axes_num=2, layout=(2, 1))
-
-        with tm.assert_produces_warning(UserWarning):
-            axes = _check_plot_works(df.height.hist, by=df.gender, layout=(3, -1))
-        self._check_axes_shape(axes, axes_num=2, layout=(3, 1))
-
-        with tm.assert_produces_warning(UserWarning):
-            axes = _check_plot_works(df.height.hist, by=df.category, layout=(4, 1))
-        self._check_axes_shape(axes, axes_num=4, layout=(4, 1))
-
-        with tm.assert_produces_warning(UserWarning):
-            axes = _check_plot_works(df.height.hist, by=df.category, layout=(2, -1))
-        self._check_axes_shape(axes, axes_num=4, layout=(2, 2))
-
-        with tm.assert_produces_warning(UserWarning):
-            axes = _check_plot_works(df.height.hist, by=df.category, layout=(3, -1))
-        self._check_axes_shape(axes, axes_num=4, layout=(3, 2))
-
-        with tm.assert_produces_warning(UserWarning):
-            axes = _check_plot_works(df.height.hist, by=df.category, layout=(-1, 4))
-        self._check_axes_shape(axes, axes_num=4, layout=(1, 4))
-
-        with tm.assert_produces_warning(UserWarning):
-            axes = _check_plot_works(df.height.hist, by=df.classroom, layout=(2, 2))
-        self._check_axes_shape(axes, axes_num=3, layout=(2, 2))
-
-        axes = df.height.hist(by=df.category, layout=(4, 2), figsize=(12, 7))
-        self._check_axes_shape(axes, axes_num=4, layout=(4, 2), figsize=(12, 7))
-
-    def test_hist_no_overlap(self):
-        from matplotlib.pyplot import gcf, subplot
-
-        x = Series(np.random.randn(2))
-        y = Series(np.random.randn(2))
-        subplot(121)
-        x.hist()
-        subplot(122)
-        y.hist()
-        fig = gcf()
-        axes = fig.axes
-        assert len(axes) == 2
-
-    def test_hist_secondary_legend(self):
-        # GH 9610
-        df = DataFrame(np.random.randn(30, 4), columns=list("abcd"))
-
-        # primary -> secondary
-        _, ax = self.plt.subplots()
-        ax = df["a"].plot.hist(legend=True, ax=ax)
-        df["b"].plot.hist(ax=ax, legend=True, secondary_y=True)
-        # both legends are dran on left ax
-        # left and right axis must be visible
-        self._check_legend_labels(ax, labels=["a", "b (right)"])
-        assert ax.get_yaxis().get_visible()
-        assert ax.right_ax.get_yaxis().get_visible()
-        tm.close()
-
-        # secondary -> secondary
-        _, ax = self.plt.subplots()
-        ax = df["a"].plot.hist(legend=True, secondary_y=True, ax=ax)
-        df["b"].plot.hist(ax=ax, legend=True, secondary_y=True)
-        # both legends are draw on left ax
-        # left axis must be invisible, right axis must be visible
-        self._check_legend_labels(ax.left_ax, labels=["a (right)", "b (right)"])
-        assert not ax.left_ax.get_yaxis().get_visible()
-        assert ax.get_yaxis().get_visible()
-        tm.close()
-
-        # secondary -> primary
-        _, ax = self.plt.subplots()
-        ax = df["a"].plot.hist(legend=True, secondary_y=True, ax=ax)
-        # right axes is returned
-        df["b"].plot.hist(ax=ax, legend=True)
-        # both legends are draw on left ax
-        # left and right axis must be visible
-        self._check_legend_labels(ax.left_ax, labels=["a (right)", "b"])
-        assert ax.left_ax.get_yaxis().get_visible()
-        assert ax.get_yaxis().get_visible()
-        tm.close()
-
     def test_df_series_secondary_legend(self):
         # GH 9779
         df = DataFrame(np.random.randn(30, 3), columns=list("abc"))
